feat: disable nginx's default server catch-all behavior
By default, nginx will route any hostname that does not match a server block to the first server block (in lexicographical order) if there is one listening on the requested port. While it can be useful, this behavior isn't something that users expect, and so we implement a hack to block it for at least ports 80 and 443 if there isn't a detected catch-all domain.

Ideally we create a server block for each port being listened to and associated domains there, but this implementation is good enough for now.