feat: add workflow to bump version in ci
Version bumps are done via git tagging, which can be annoying to do locally as I would need to lookup the last release. Instead, automate the actual workflow so that it can be bumped in CI.

Ideally version bumps are done on PR merge and automatically computed, but this workflow is fine for now.